improved_insights/requirements.txt
Flask==2.1.0
flask_cors
langchain
llama_index
python-dotenv
gunicorn
